Move ToggleElement to ui
